abstract strategy modified to remember the last gamestate
