+++ src/main/java/idunno/spacescavanger/ClientEndpoint.java
@@ -2,9 +2,13 @@ package idunno.spacescavanger;
 
 import static idunno.spacescavanger.dto.GameStatus.ABORTED;
 import static idunno.spacescavanger.dto.GameStatus.ENDED;
+import static idunno.spacescavanger.strategy.Comparators.compareByDistance;
 import static java.util.Optional.empty;
+import static java.util.function.Function.identity;
 
+import java.util.Map;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import javax.websocket.Endpoint;
 import javax.websocket.EndpointConfig;
@@ -15,6 +19,9 @@ import idunno.spacescavanger.dto.Game;
 import idunno.spacescavanger.dto.GameResponse;
 import idunno.spacescavanger.dto.GameState;
 import idunno.spacescavanger.dto.GameStatus;
+import idunno.spacescavanger.dto.Meteorite;
+import idunno.spacescavanger.dto.Ship;
+import idunno.spacescavanger.strategy.Position;
 import idunno.spacescavanger.strategy.Strategy;
 
 public class ClientEndpoint extends Endpoint implements MessageHandler.Whole<String> {
@@ -54,11 +61,43 @@ public class ClientEndpoint extends Endpoint implements MessageHandler.Whole<Str
 	}
 
 	private void initStrategy(String message) {
-		strategy = Optional.of(new Strategy.NoStrategy(converter.toObject(message, Game.class)));
+		strategy = Optional.of(new NoStrategy(converter.toObject(message, Game.class)));
 	}
 
 	public void sendMessage(String message) {
 		session.getAsyncRemote()
 				.sendText(message);
 	}
+
+	// TODO delete this.
+	public static class NoStrategy extends Strategy {
+
+		public NoStrategy(Game game) {
+			super(game);
+		}
+
+		@Override
+		public GameResponse suggestMove(GameState lastState, GameState currentState) {
+			return suggestFirstMove(currentState);
+		}
+
+		@Override
+		protected GameResponse suggestFirstMove(GameState currentState) {
+			Map<String, Ship> shipsByOwner = currentState.getShipStates()
+					.stream()
+					.collect(Collectors.toMap(Ship::getOwner, identity()));
+			Optional<Position> min = currentState.getMeteoriteStates()
+					.stream()
+					.map(Meteorite::getPosition)
+					.min(compareByDistance(shipsByOwner.get("idunno")
+							.getPosition()));
+			return GameResponse.builder()
+					.withShipMoveToPosition(min.orElse(new Position(100, 100)))
+					.withRocketMoveToPosition(shipsByOwner.get("bot1")
+							.getPosition())
+					.build();
+		}
+
+	}
+
 }
