+++ src/main/java/idunno/spacescavanger/strategy/Comparators.java
@@ -6,7 +6,7 @@ import java.util.Comparator;
 
 public class Comparators {
 
-	static Comparator<Position> compareByDistance(Position from) {
+	public static Comparator<Position> compareByDistance(Position from) {
 		return (first, second) -> compare(from.distance(first), from.distance(second));
 	}
 }
